[core-amqp] Move to shared rollup config
- Use the shared rollup config from dev-tool for product.

  while working on this, move `StandardAbortMessage` from error.ts to
  util/constants.ts to break the circular dependency of errors -> utils
  -> errors.
