+++ sdk/core/core-amqp/rollup.config.js
@@ -1,16 +1,5 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+import { makeConfig } from "@azure/dev-tool/shared-config/rollup";
 
-import * as base from "./rollup.base.config";
-
-const inputs = [];
-
-if (!process.env.ONLY_BROWSER) {
-  inputs.push(base.nodeConfig());
-}
-
-if (!process.env.ONLY_NODE) {
-  inputs.push(base.browserConfig());
-}
-
-export default inputs;
+export default makeConfig(require("./package.json"));
