+++ sdk/core/core-amqp/src/util/utils.ts
@@ -3,7 +3,7 @@
 
 import { AbortError, AbortSignalLike } from "@azure/abort-controller";
 import { CancellableAsyncLock, CancellableAsyncLockImpl } from "./lock";
-import { StandardAbortMessage } from "../errors";
+import { StandardAbortMessage } from "./constants";
 import { WebSocketImpl } from "rhea-promise";
 import { isDefined } from "./typeGuards";
